Unfinished implementation of two-way binding for macOS
Cocoa doesn't use UITextInput, which UIKit does, and which
we already have special consideration for in regard to IME
support. There might be other things to consider as well,
so before just throwing in simple support for Cocoa text
views, more manual testing should be done for various
languages and behaviors.